Module rust_im::traffic_detection

Structs

SearchVector
Used to keep track of the area of an image that will be searched.

ThreeLaneBlob
The structure used to perform simplified blob analysis on images from the first ~30 sec from this video. Technically, the library could be used on any video that has 3 lanes of traffic and a similar overhead camera angle, but your mileage may very as we had little luck finding other ones similar to this. The inspiration for this idea came from this 2015 paper by Ashwini B. and Yvaraju B N, though the actual blob analysis and most of the implementation details used in this struct are much simpler than the technique they used.

VehicleBlock

Functions

class_demo

count_example