Struct rust_im::traffic_detection::SearchVector

pub struct SearchVector { /* fields omitted */ }

Used to keep track of the area of an image that will be searched.

For the purposes of this project, that will be a highway lane.

Examples

To search a box of width 20 pixels that runs from some pixel (250, 50) to (45, 615), we could create a search vector like so:

let left_sv = let left_sv = SearchVector::new((250, 50), (45, 615), 10);

The SearchVector will internally keep track of slope to ensure that the area is searched for blobs correctly, meaning that SearchVectors can be angled. It is important when using this struct in conjuntion with ThreeLaneBlob that they have a slope with absolute value > 1. The higher the slopes absolute value, the better the results will be due to the underlying implementation of that struct.

Methods

impl SearchVector

pub fn new(start: (u32, u32), end: (u32, u32), width_over_two: u32) -> Self