Struct rust_im::traffic_detection::ThreeLaneBlob

pub struct ThreeLaneBlob { /* fields omitted */ }

The structure used to perform simplified blob analysis on images from the first ~30 sec from this video. Technically, the library could be used on any video that has 3 lanes of traffic and a similar overhead camera angle, but your mileage may very as we had little luck finding other ones similar to this. The inspiration for this idea came from this 2015 paper by Ashwini B. and Yvaraju B N, though the actual blob analysis and most of the implementation details used in this struct are much simpler than the technique they used.

Methods

impl ThreeLaneBlob

pub fn new(
im: DynamicImage,
left: SearchVector,
middle: SearchVector,
right: SearchVector
) -> Self

Examples

You can create a new ThreeLaneBlob using a screenshot from the video like so:

let my_im = image::open(filename_of_screenshot).unwrap();
let left_sv = SearchVector::new((250, 50), (45, 615), 10);
let mid_sv = SearchVector::new((300, 50), (300, 650), 10);
let right_sv = SearchVector::new((350, 50), (570, 650), 10);
let my_blob_test = ThreeLaneBlob::new(my_im, left_sv, mid_sv, right_sv);

The left parameter is the area that will be searched in the left lane. The middle and right parameters denote the areas that will be searched in their respective lanes as well. For more info on how these work, see SearchVector.

An important thing to note is that the SearchVector for the middle lane can handle having a an infinite slope (see mid_sv above), while the SearchVectors for the other lanes cannot. This is meant to ensure that the lanes are always centered correctly, if any of them must have a zero slope.

pub fn process_directory_of_color_ims(
&mut self,
input_directory: &str,
output_directory: &str,
sobel_threshold: u8
)

Allows for the processing of multiple images in sequence (however many are in input_directory.

Examples

Since a ThreeLaneBlob must be already instantiated for this method to be called, you just need to make one with the first image of the directory to be processed, then run this method on the ThreeLaneBlob.

let new_blob_test = ThreeLaneBlob::new(.......);
new_blob_test.process_directory_of_color_ims("test_in\\", "test_out\\", 70);

pub fn process_sobeled_image(&mut self, output_directory: &str)

Used to process an image that has already had a sobel filter applied to it.

pub fn process_color_image(
&mut self,
output_directory: &str,
sobel_output_name: &str,
outlined_output_name: &str,
sobel_threshold: u8
)

Used to process an image that has not yet had a sobel filter applied to it.

pub fn get_car_count(&self) -> usize

Returns the number of cars/blobs that it has found in the image.